backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
openai
google-generativeai
clerk-backend-api
svix
